refactor: extract buildSavedAttachment in attachments.ts helpers
- consider to define a "well-defined" module instead of making a "helper"
