+++ src/helpers/attachments.test.tsx
@@ -5,7 +5,12 @@
  */
 // noinspection HtmlRequiredLangAttribute
 
-import { filterAttachmentsParts, getReferredContentIds } from 'helpers/attachments';
+import { generateMessage } from '../tests/generators/generateMessage';
+import {
+	buildSavedAttachments,
+	filterAttachmentsParts,
+	getReferredContentIds
+} from 'helpers/attachments';
 import type { MailMessagePart } from 'types/index.d';
 
 describe('attachments', () => {
@@ -93,6 +98,175 @@ describe('attachments', () => {
 
 			expect(result).toHaveLength(1);
 			expect(result[0].filename).toBe('logo.jpg');
+		});
+	});
+
+	describe('buildSavedAttachments', () => {
+		it('should return an empty array when there are no parts', () => {
+			const message = generateMessage({ folderId: '2' });
+			message.parts = [];
+			const result = buildSavedAttachments(message);
+			expect(result).toEqual([]);
+		});
+		it('should set attachment as not inline if disposition is not inline', () => {
+			const message = generateMessage({ folderId: '2' });
+			message.parts = [
+				{
+					contentType: 'image/png',
+					filename: 'img.png',
+					disposition: 'attachment',
+					name: '2.2',
+					size: 1234,
+					ci: '<abc123@zimbra>'
+				}
+			];
+			const result = buildSavedAttachments(message);
+
+			expect(result[0].isInline).toBeFalsy();
+		});
+		it('should mark attachment with contentId and type image/* as NOT inline if NOT referenced in the body ', () => {
+			const message = generateMessage({ folderId: '2' });
+			message.parts = [
+				{
+					contentType: 'text/html',
+					size: 0,
+					name: 'HTML body',
+					content: 'This is my inline image: <a href="wrongCIDReference:<abc123@zimbra>"/>'
+				},
+				{
+					contentType: 'image/png',
+					filename: 'img.png',
+					name: '2.2',
+					size: 1234,
+					disposition: undefined,
+					ci: '<abc123@zimbra>'
+				}
+			];
+
+			const result = buildSavedAttachments(message);
+
+			expect(result[0].isInline).toBeFalsy();
+		});
+		it('should mark attachment with contentId and type image/* as inline if referenced in the body ', () => {
+			const message = generateMessage({ folderId: '2' });
+			message.parts = [
+				{
+					contentType: 'text/html',
+					size: 0,
+					name: 'HTML body',
+					content: 'This is my inline image: <a href="cid:<abc123@zimbra>"/>'
+				},
+				{
+					contentType: 'image/png',
+					filename: 'img.png',
+					name: '2.2',
+					size: 1234,
+					disposition: undefined,
+					ci: '<abc123@zimbra>'
+				}
+			];
+
+			const result = buildSavedAttachments(message);
+
+			expect(result[0]).toMatchObject({
+				isInline: true,
+				contentId: 'abc123@zimbra',
+				partName: '2.2',
+				contentType: 'image/png',
+				filename: 'img.png',
+				messageId: message.id
+			});
+		});
+
+		it('should mark part with disposition "inline" as inline even if not an image', () => {
+			const message = generateMessage({ folderId: '2' });
+			message.parts = [
+				{
+					contentType: 'application/pdf',
+					disposition: 'inline',
+					filename: 'doc.pdf',
+					name: '2.3',
+					size: 2048
+				}
+			];
+
+			const result = buildSavedAttachments(message);
+			expect(result[0].isInline).toBe(true);
+		});
+
+		it('should not mark as inline when contentId is missing and disposition is not "inline"', () => {
+			const message = generateMessage({ folderId: '2' });
+			message.parts = [
+				{
+					contentType: 'application/pdf',
+					filename: 'doc.pdf',
+					name: '2.4',
+					size: 512
+				}
+			];
+
+			const result = buildSavedAttachments(message);
+			expect(result[0].isInline).toBe(false);
+		});
+
+		it('should extract inner contentId from brackets', () => {
+			const message = generateMessage({ folderId: '2' });
+			message.parts = [
+				{
+					contentType: 'image/jpeg',
+					ci: '<image123@crb>',
+					name: '2.5',
+					size: 200
+				}
+			];
+
+			const result = buildSavedAttachments(message);
+			expect(result[0].contentId).toBe('image123@crb');
+		});
+
+		it('should leave contentId undefined if ci is not present', () => {
+			const message = generateMessage({ folderId: '2' });
+			message.parts = [
+				{
+					contentType: 'image/jpeg',
+					name: '2.6',
+					size: 300
+				}
+			];
+
+			const result = buildSavedAttachments(message);
+			expect(result[0].contentId).toBeUndefined();
+		});
+
+		it('should correctly build a SavedAttachment for a part with inline CID and text/html content', () => {
+			const message = generateMessage({ folderId: '2' });
+
+			message.parts = [
+				{
+					contentType: 'text/html',
+					content: `<html><body>
+				<img src="cid:65766eee-4439-438c-a375-1ac111ed1a07@zimbra" />
+				<p>Hello, this is a test email with inline image.</p>
+			</body></html>`,
+					size: 999,
+					name: '2.2',
+					ci: '<65766eee-4439-438c-a375-1ac111ed1a07@zimbra>'
+				}
+			];
+
+			const result = buildSavedAttachments(message);
+
+			expect(result).toEqual([
+				{
+					messageId: message.id,
+					isInline: true, // because ci is present and contentType is text/html
+					contentId: '65766eee-4439-438c-a375-1ac111ed1a07@zimbra',
+					filename: '', // no filename provided
+					partName: '2.2',
+					contentType: 'text/html',
+					size: 999
+				}
+			]);
 		});
 	});
 });
