+++ src/views/app/detail-panel/edit/parts/editor-paste-handler.ts
@@ -8,12 +8,9 @@ import type { Editor } from 'tinymce';
 import { v4 as uuid } from 'uuid';
 
 import { uploadFileApi } from 'api/upload-file-api';
-import { composeAttachmentDownloadUrl } from 'helpers/attachments';
+import { buildSavedAttachments, composeAttachmentDownloadUrl } from 'helpers/attachments';
 import { normalizeMailMessageFromSoap } from 'normalizations/normalize-message';
-import {
-	buildSavedAttachments,
-	composeCidUrlFromContentId
-} from 'store/editor/editor-transformations';
+import { composeCidUrlFromContentId } from 'store/editor/editor-transformations';
 import { getSavedInlineAttachmentByContentId } from 'store/editor/editor-utils';
 import { getEditor, useEditorsStore } from 'store/editor/index';
 import { saveDraftEmailStoreAction } from 'store/emails/actions/save-draft-action';
